Fix scrolling in Sites tab with FlatLaf L&F
Check the scrollable unit increment of the Sites tree if the Contexts
tree reports no more increments, it might be that the view is almost in
the Sites tree but not yet in it, which would cause the Contexts tree to
report no more increments while it might still have more for the Sites
tree.

Fix #6520.

Signed-off-by: thc202 <thc202@gmail.com>